@@ -1,6 +1,6 @@
 {
   "name": "elemental",
-  "version": "0.4.5",
+  "version": "0.5.6",
   "description": "React UI Framework",
   "main": "lib/Elemental.js",
   "author": "Jed Watson",
@@ -10,25 +10,30 @@
     "url": "https://github.com/elementalui/elemental.git"
   },
   "dependencies": {
+    "ally.js": "^1.0.1",
     "blacklist": "^1.1.2",
-    "classnames": "^2.1.3",
-    "react-interval": "^1.0.6"
+    "classnames": "^2.2.0"
   },
   "devDependencies": {
     "babel": "^5.8.23",
     "babel-eslint": "^4.1.3",
     "chai": "^3.2.0",
-    "eslint": "^1.4.3",
-    "eslint-plugin-react": "^3.4.1",
-    "gulp": "^3.9.0",
     "jsdom": "^2.0.0",
     "mocha": "^2.3.1",
-    "react": "^0.13.3",
-    "react-component-gulp-tasks": "^0.7.1",
-    "react-router": "^0.13.3"
+    "eslint": "^1.6.0",
+    "eslint-plugin-react": "^3.5.1",
+    "gulp": "^3.9.0",
+    "history": "^1.12.3",
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0",
+    "react-addons-css-transition-group": "^0.14.0",
+    "react-component-gulp-tasks": "^0.7.6",
+    "react-router": "1.0.0-rc3"
   },
   "peerDependencies": {
-    "react": "^0.13.0 || ^0.14.0-beta"
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0",
+    "react-addons-css-transition-group": "^0.14.0"
   },
   "browserify-shim": {
     "react": "global:React"
@@ -42,6 +47,7 @@
     "publish:npm": "gulp publish:npm",
     "publish:site": "gulp publish:examples",
     "publish:tag": "gulp publish:tag",
+    "publish:bower": "./scripts/bower.sh",
     "release": "gulp release",
     "site": "gulp dev:server",
     "start": "gulp dev",
