@@ -1,22 +1,22 @@
-var blacklist = require('blacklist');
-var classNames = require('classnames');
-var React = require('react/addons');
+const blacklist = require('blacklist');
+const classNames = require('classnames');
+const React = require('react');
 
-var Checkbox = React.createClass({
+const Checkbox = React.createClass({
 	propTypes: {
 		className: React.PropTypes.string,
 		disabled: React.PropTypes.bool,
-		focusOnMount: React.PropTypes.bool,
+		autofocus: React.PropTypes.bool,
 		indeterminate: React.PropTypes.bool,
 		inline: React.PropTypes.bool,
 		label: React.PropTypes.string,
 		style: React.PropTypes.object,
-		title: React.PropTypes.string
+		title: React.PropTypes.string,
 	},
 
 	componentDidMount () {
-		if (this.props.focusOnMount) {
-			this.refs.target.getDOMNode().focus();
+		if (this.props.autofocus) {
+			this.refs.target.focus();
 		}
 		this.setIndeterminate(this.props.indeterminate);
 	},
@@ -26,16 +26,15 @@ var Checkbox = React.createClass({
 	},
 
 	setIndeterminate (value) {
-		this.refs.target.getDOMNode().indeterminate = value;
+		this.refs.target.indeterminate = value;
 	},
 
 	render() {
-		var componentClass = classNames('Checkbox', {
+		let componentClass = classNames('Checkbox', {
 			'Checkbox--disabled': this.props.disabled,
-			'Checkbox--inline': this.props.inline
+			'Checkbox--inline': this.props.inline,
 		}, this.props.className);
-		var props = blacklist(this.props, 'className', 'label', 'style', 'title');
-
+		let props = blacklist(this.props, 'className', 'label', 'style', 'title');
 		return (
 			<label className={componentClass} style={this.props.style} title={this.props.title}>
 				<input ref="target" type="checkbox" className="Checkbox__input" {...props} />
