@@ -1,31 +1,27 @@
-var React = require('react/addons');
-var ReactCSSTransitionGroup = React.addons.CSSTransitionGroup;
-var blacklist = require('blacklist');
-var classNames = require('classnames');
-var ReactInterval = require('react-interval');
-var { shouldComponentUpdate } = require('react/lib/ReactComponentWithPureRenderMixin');
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Transition from 'react-addons-css-transition-group';
+import blacklist from 'blacklist';
+import classNames from 'classnames';
+import ally from 'ally.js';
 
-const ESC_KEYCODE = 27;
-
-const safeTop = element => {
-	const { innerHeight, pageYOffset, getComputedStyle } = window;
-	const { clientHeight, offsetTop } = element;
-	const { marginBottom, marginTop } = getComputedStyle(element);
-	const [mTop, mBottom] = [parseInt(marginTop, 10), parseInt(marginBottom, 10)];
-	const height = clientHeight + mTop + mBottom;
-
-	if ((offsetTop - mTop) < pageYOffset && offsetTop - mTop + height > pageYOffset + innerHeight) {
-		// Scrolling within modal content, don't move
-		return false;
+const TransitionPortal = React.createClass({
+	displayName: 'TransitionPortal',
+	portalElement: null,
+	render: () => null,
+	componentDidMount() {
+		let p = document.createElement('div');
+		document.body.appendChild(p);
+		this.portalElement = p;
+		this.componentDidUpdate();
+	},
+	componentWillUnmount() {
+		document.body.removeChild(this.portalElement);
+	},
+	componentDidUpdate() {
+		ReactDOM.render(<Transition {...this.props}>{this.props.children}</Transition>, this.portalElement);
 	}
-	if (offsetTop - mTop < pageYOffset && height > innerHeight) {
-		// Stick to the window bottom
-		return pageYOffset + innerHeight - height;
-	} else {
-		// Stick to the window top
-		return pageYOffset;
-	}
-};
+});
 
 module.exports = React.createClass({
 	displayName: 'Modal',
@@ -34,64 +30,125 @@ module.exports = React.createClass({
 		className: React.PropTypes.string,
 		isOpen: React.PropTypes.bool,
 		onCancel: React.PropTypes.func,
-		top: React.PropTypes.number
+		width: React.PropTypes.oneOfType([
+			React.PropTypes.oneOf(['small', 'medium', 'large']),
+			React.PropTypes.number,
+		]),
 	},
-	getInitialState() {
-		return { top: typeof this.props.top !== 'undefined' ? this.props.top : window.pageYOffset };
+	getDefaultProps () {
+		return {
+			width: 'medium'
+		};
 	},
 	componentWillReceiveProps: function(nextProps) {
 		if (nextProps.isOpen) {
-			window.addEventListener('keydown', this.handleKeyDown);
+			setTimeout(() => this.handleAccessibility());
+			document.body.style.overflow = 'hidden';
 		} else {
-			window.removeEventListener('keydown', this.handleKeyDown);
+			setTimeout(() => this.removeAccessibilityHandlers());
 		}
 	},
-	shouldComponentUpdate,
-	handleKeyDown (e) {
-		if (e.keyCode === ESC_KEYCODE) {
-			this.props.onCancel();
-		}
+	handleAccessibility () {
+		// Remember the element that was focused before we opened the modal
+		// so we can return focus to it once we close the modal.
+		this.focusedElementBeforeModalOpened = document.activeElement;
+
+		// We're using a transition to reveal the modal,
+		// so wait until the element is visible, before
+		// finding the first keyboard focusable element
+		// and passing focus to it, otherwise the browser
+		// might scroll the document to reveal the element
+		// receiving focus
+		ally.when.visibleArea({
+			context: this.modalElement,
+			callback: function(context) {
+				// the modal is visible on screen, so find the first
+				// keyboard focusable element (giving any element with
+				// autofocus attribute precendence). If the modal does
+				// not contain any keyboard focusabe elements, focus will
+				// be given to the modal itself.
+				var element = ally.query.firstTabbable({
+					context: context,
+					defaultToContext: true,
+				});
+				element.focus();
+			},
+		});
+
+		// Make sure that no element outside of the modal
+		// can be interacted with while the modal is visible.
+		this.disabledHandle = ally.maintain.disabled({
+			filter: this.modalElement,
+		});
+
+		// Make sure that no element outside of the modal
+		// is exposed via the Accessibility Tree, to prevent
+		// screen readers from navigating to content it shouldn't
+		// be seeing while the modal is open.
+		this.hiddenHandle = ally.maintain.hidden({
+			filter: this.modalElement,
+		});
+
+		// React to escape keys as mandated by ARIA Practices
+		this.keyHandle = ally.when.key({
+			escape: this.handleClose,
+		});
 	},
-	updateTop() {
-		const top = safeTop(React.findDOMNode(this.refs.dialog));
-		if (top !== false) {
-			this.setState({ top });
-		}
+	removeAccessibilityHandlers () {
+		// undo listening to keyboard
+		this.keyHandle && this.keyHandle.disengage();
+
+		// undo hiding elements outside of the modal
+		this.hiddenHandle && this.hiddenHandle.disengage();
+
+		// undo disabling elements outside of the modal
+		this.disabledHandle && this.disabledHandle.disengage();
+
+		// return focus to where it was before we opened the modal
+		this.focusedElementBeforeModalOpened && this.focusedElementBeforeModalOpened.focus();
+	},
+	handleModalClick (event) {
+		if (event.target.dataset.modal) this.handleClose();
+	},
+	handleClose () {
+		document.body.style.overflow = null;
+		this.props.onCancel();
 	},
 	renderDialog() {
-		if (!this.props.isOpen) return null;
+		if (!this.props.isOpen) return;
+
+		let dialogClassname = classNames('Modal-dialog', (this.props.width && isNaN(this.props.width)) ? (
+			'Modal-dialog--' + this.props.width
+		) : null);
 
 		return (
-			<div ref="dialog" className="Modal-dialog" style={{ top: this.state.top }}>
-				<ReactInterval timeout={200} enabled={typeof this.props.top === 'undefined'}
-					callback={this.updateTop} />
-				<div className="Modal-content">
+			<div className={dialogClassname} style={(this.props.width && !isNaN(this.props.width)) ? { width: this.props.width + 20 } : null}>
+				<div ref={ref => { this.modalElement = ref; }} className="Modal-content">
 					{this.props.children}
 				</div>
 			</div>
 		);
 	},
 	renderBackdrop() {
-		if (!this.props.isOpen) return null;
+		if (!this.props.isOpen) return;
 
-		return <div className="Modal-backdrop" onClick={this.props.backdropClosesModal ? this.props.onCancel : null} />;
+		return <div className="Modal-backdrop" onClick={this.props.backdropClosesModal ? this.handleClose : null} />;
 	},
 	render() {
-		// classes
-		var className = classNames('Modal', this.props.className);
+		var className = classNames('Modal', {
+			'is-open': this.props.isOpen
+		}, this.props.className);
 
-		// props
-		var props = blacklist(this.props, 'backdropClosesModal', 'className', 'headerHasCloseButton', 'headerTitle', 'isOpen');
-		props.className = className;
+		var props = blacklist(this.props, 'backdropClosesModal', 'className', 'isOpen', 'onCancel');
 
 		return (
-			<div {...props}>
-				<ReactCSSTransitionGroup transitionName="Modal-dialog" component="div">
+			<div>
+				<TransitionPortal {...props} data-modal="true" className={className} onClick={this.handleModalClick} transitionName="Modal-dialog" transitionEnterTimeout={260} transitionLeaveTimeout={140} component="div">
 					{this.renderDialog()}
-				</ReactCSSTransitionGroup>
-				<ReactCSSTransitionGroup transitionName="Modal-background" component="div">
+				</TransitionPortal>
+				<TransitionPortal transitionName="Modal-background" transitionEnterTimeout={140} transitionLeaveTimeout={240} component="div">
 					{this.renderBackdrop()}
-				</ReactCSSTransitionGroup>
+				</TransitionPortal>
 			</div>
 		);
 	}
