@@ -1,4 +1,4 @@
-var React = require('react/addons');
+var React = require('react');
 var classNames = require('classnames');
 var blacklist = require('blacklist');
 
@@ -30,6 +30,7 @@ module.exports = React.createClass({
 		block: React.PropTypes.bool,
 		className: React.PropTypes.string,
 		href: React.PropTypes.string,
+		component: React.PropTypes.node,
 		isActive: React.PropTypes.bool,
 		size: React.PropTypes.oneOf(BUTTON_SIZES),
 		submit: React.PropTypes.bool,
@@ -54,21 +55,21 @@ module.exports = React.createClass({
 		);
 
 		// props
-		var props = blacklist(this.props, 'type', 'size', 'className');
+		var props = blacklist(this.props, 'type', 'size', 'component', 'className');
 		props.className = componentClass;
+
+		if (this.props.component) {
+			return React.cloneElement(this.props.component, props);
+		}
 
 		var tag = 'button';
 		props.type = this.props.submit ? 'submit' : 'button';
 
 		if (props.href) {
 			tag = 'a';
-			props.type = null;
+			delete props.type;
 		}
 
-		return React.createElement(
-			tag,
-			props,
-			this.props.children
-		);
+		return React.createElement(tag, props, this.props.children);
 	}
 });
