@@ -1,4 +1,4 @@
-var React = require('react/addons');
+var React = require('react');
 var blacklist = require('blacklist');
 
 var Button = require('./Button');
@@ -25,9 +25,13 @@ module.exports = React.createClass({
 			loading: false
 		};
 	},
-
+        componentDidMount () {
+               this.refs.fileInput.addEventListener('click', function () {
+                       this.value = '';
+               },   false);
+        },
 	triggerFileBrowser() {
-		this.refs.fileInput.getDOMNode().click();
+		this.refs.fileInput.click();
 	},
 	handleChange(e) {
 		var self = this;
